backendcode/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.0
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.6.2
boto3==1.24.86
botocore==1.27.86
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
chardet==5.0.0
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.6.2
cli-helpers==2.2.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
contourpy==1.0.7
cryptography==41.0.3
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==3.2.9
django-allauth==0.54.0
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap-pagination==1.7.1
django-bootstrap3==22.1
django-cors-headers==4.2.0
django-filter==23.2
django-impersonate==1.8.2
django-ipware==4.0.2
django-mysql==4.7.1
django-paginator-rawqueryset @ git+https://github.com/d-b/django-paginator-rawqueryset.git@e3a02d56399640330a7d903216f32f6d562897bb
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
document==1.0
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.95.2
fastapi-utils==0.2.1
fasteners==0.14.1
filelock==3.6.0
fonttools @ file:///home/studio/Filehandling12june/Filehandling/dependencies/fonttools-4.39.4-py3-none-any.whl
frozenlist==1.3.3
future==0.18.2
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
gyp==0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.20.2
hyperframe==6.0.1
idna @ file:///home/studio/Filehandling12june/Filehandling/dependencies/idna-3.4-py3-none-any.whl
importlib-metadata==5.0.0
jeepney==0.7.1
Jinja2 @ file:///home/studio/Filehandling12june/Filehandling/dependencies/Jinja2-3.1.2-py3-none-any.whl
jmespath==1.0.1
kaitaistruct==0.10
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe @ file:///home/studio/Filehandling12june/Filehandling/dependencies/MarkupSafe-2.1.2-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
matplotlib==3.7.1
monotonic==1.6
more-itertools==8.14.0
mpld3 @ file:///home/studio/Filehandling12june/Filehandling/dependencies/mpld3-0.5.9-py3-none-any.whl
mssql-django==1.1.3
multidict==6.0.4
mycli==1.24.3
mysql-connector==2.2.9
mysqlclient==2.1.1
netifaces==0.11.0
networkx==3.1
numpy==1.25.2
oauth2==1.9.0.post1
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
ordereddict==1.1
outcome==1.2.0
packaging==21.3
panda @ file:///home/studio/Filehandling12june/Filehandling/dependencies/panda-0.3.1.tar.gz
pandas==2.1.0
pandas-datareader @ file:///home/studio/Filehandling12june/Filehandling/dependencies/pandas_datareader-0.10.0-py3-none-any.whl
paramiko==2.9.3
pexpect==4.8.0
Pillow @ file:///home/studio/Filehandling12june/Filehandling/dependencies/Pillow-9.5.0-cp310-cp310-manylinux_2_28_x86_64.whl
pip-review==1.2.0
platformdirs==2.5.1
plotly @ file:///home/studio/Filehandling12june/Filehandling/dependencies/plotly-5.14.1-py2.py3-none-any.whl
prompt-toolkit==3.0.31
protobuf==4.24.2
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.8
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.5.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyodbc==4.0.34
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pyRFC3339==1.1
PySocks==1.7.1
pyTelegramBotAPI==4.13.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-memcached==1.59
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2022.4
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
SecretStorage==3.3.1
selenium==4.15.1
selenium-wire==5.1.0
six @ file:///home/studio/Filehandling12june/Filehandling/dependencies/six-1.16.0-py2.py3-none-any.whl
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.48
sqlparse==0.4.3
starlette==0.27.0
systemd-python==234
tabulate==0.8.9
telebot==0.0.5
tenacity @ file:///home/studio/Filehandling12june/Filehandling/dependencies/tenacity-8.2.2-py3-none-any.whl
terminaltables==3.1.0
trio==0.22.2
trio-websocket==0.10.4
twilio==7.14.1
typing_extensions==4.6.2
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.7
uvicorn==0.22.0
vine==5.0.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
WooCommerce==3.0.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==3.8.1
zstandard==0.21.0
